moved metrics to console
