+++ excalibur/console/__init__.py
@@ -2,7 +2,7 @@
 
 # -- IMPORTS -- ------------------------------------------------------
 import dawgie
-import excalibur.metrics.bot
+import excalibur.console.bot
 
 from importlib import import_module as fetch  # avoid cyclic-import
 
@@ -12,15 +12,15 @@ DAWGIE_IGNORE = False
 
 
 def analysis(prefix: str, ps_hint: int = 0, runid: int = -1):
-    '''metrics are global or an aspect'''
-    return excalibur.metrics.bot.AnalysisTeam(prefix, ps_hint, runid)
+    '''console is global or an aspect'''
+    return excalibur.console.bot.AnalysisTeam(prefix, ps_hint, runid)
 
 
 def events():
     return [
         dawgie.schedule(
             analysis,
-            fetch('excalibur.metrics.algorithms').Performance(),
+            fetch('excalibur.console.algorithms').Performance(),
             boot=True,
         )
     ]
