+++ excalibur/console/algorithms.py
