+++ excalibur/console/bot.py
